GammaPredictor abstraction
