+++ src/runner/predictors/__init__.py
@@ -1,10 +1,12 @@
 from .base_predictor import BasePredictor
+from .gamma_predictor import GammaPredictor
 from .acdc_seg_predictor import AcdcSegPredictor
 from .lits_seg_predictor import LitsSegPredictor
 from .brats17_seg_predictor import Brats17SegPredictor
 
 __all__ = [
     'BasePredictor',
+    'GammaPredictor',
     'AcdcSegPredictor',
     'LitsSegPredictor',
     'Brats17SegPredictor',
