+++ src/runner/predictors/lits_seg_predictor.py
@@ -1,77 +1,16 @@
-import logging
 import torch
 import nibabel as nib
-import matplotlib.pyplot as plt
 import torch.nn.functional as F
 
-from src.runner.predictors import BasePredictor
-from src.runner.predictors.utils import clip_gamma, get_gamma_percentange
-
-LOGGER = logging.getLogger(__name__.split('.')[-1])
+from src.runner.predictors import GammaPredictor
 
 
-class LitsSegPredictor(BasePredictor):
+class LitsSegPredictor(GammaPredictor):
     """The LiTS predictor for the segmentation task.
-    Args:
-        saved_pred (bool): Whether to save the prediction (default: False).
-        plot_gamma_performance_curve (bool): Whether to plot gamma vs performance figure (default: False).
-        gamma_thresholds (sequence): The thresholds to clip gamma value (default: None).
-            Note that this argument is only valid when plot_gamma_performance_curve is True.
     """
 
-    def __init__(self, saved_pred=False, plot_gamma_performance_curve=False, gamma_thresholds=None, **kwargs):
+    def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        if self.test_dataloader.batch_size != 1:
-            raise ValueError(f'The testing batch size should be 1. Got {self.test_dataloader.batch_size}.')
-
-        self.saved_pred = saved_pred
-        self.output_dir = self.saved_dir / 'prediction'
-        if not self.output_dir.is_dir():
-            self.output_dir.mkdir(parents=True)
-        self.plot_gamma_performance_curve = plot_gamma_performance_curve
-        gamma_thresholds = set(gamma_thresholds)
-        gamma_thresholds.add(0)
-        self.gamma_thresholds = sorted(gamma_thresholds)
-
-    def predict(self):
-        if self.plot_gamma_performance_curve is False:
-            super().predict()
-        else:
-            state_dict = self.net.state_dict()
-            test_logs = {}
-            percentages = []
-            for gamma_threshold in self.gamma_thresholds:
-                tmp_state_dict = clip_gamma(state_dict, gamma_threshold)
-                self.net.load_state_dict(tmp_state_dict)
-                print()
-                LOGGER.info(f'Gamma threshold: {gamma_threshold}')
-                test_log = super().predict()
-                if test_logs:
-                    for key in test_log.keys():
-                        test_logs[key].append(test_log[key])
-                else:
-                    for key in test_log.keys():
-                        test_logs[key] = [test_log[key]]
-                percentage = get_gamma_percentange(state_dict, gamma_threshold)
-                percentages.append(f'{percentage: .2f}%')
-            for key in test_logs.keys():
-                fig = plt.figure(figsize=(10, 8))
-                plt.axhline(y=test_logs[key][0], label='No cliping', color='black', linestyle='--')
-                plt.axhspan(test_logs[key][0] * 0.98, test_logs[key][0] * 1.02, color='black', alpha=0.1)
-                plt.plot(self.gamma_thresholds[1:], test_logs[key][1:], color='blue', marker='o')
-                for i, percentage in enumerate(percentages[1:], start=1):
-                    plt.annotate(
-                        percentage,
-                        xy=(self.gamma_thresholds[i], test_logs[key][i]),
-                        xytext=(-15, -15),
-                        textcoords='offset points',
-                        ha='center'
-                    )
-                plt.xlabel('Gamma threshold')
-                plt.ylabel(key)
-                plt.legend(loc='lower left')
-                figure_path = self.saved_dir / f'{key}.png'
-                fig.savefig(figure_path.as_posix())
 
     def _test_step(self, batch):
         if self.test_dataloader.dataset.csv_name == 'testing.csv':
@@ -97,6 +36,7 @@ class LitsSegPredictor(BasePredictor):
             dice = self.metric_fns.dice(F.softmax(output, dim=1), target)
 
         if self.saved_pred:
+            output_dir = self.saved_pred / 'prediction'
             (affine,), (header,), (name,) = batch['affine'], batch['header'], batch['name']
             _, pred = F.softmax(output, dim=1).max(dim=1)
             pred = pred.squeeze(dim=0).permute(1, 2, 0).contiguous()
@@ -106,7 +46,7 @@ class LitsSegPredictor(BasePredictor):
                     affine.numpy(),
                     header
                 ),
-                (self.output_dir / name).as_posix()
+                (output_dir / name).as_posix()
             )
         return {
             'loss': loss,
@@ -115,7 +55,6 @@ class LitsSegPredictor(BasePredictor):
                 'DiceLoss': dice_loss
             },
             'metrics': {
-                'DiceLiver': dice[1],
-                'DiceLesion': dice[2]
+                'DiceLiver': dice[1]
             }
         }
